requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Cryptography
cryptography>=41.0.0

# HTTP and WebSocket
requests>=2.31.0
websockets>=12.0

# Polymarket client
py-clob-client>=0.20.0

# Ethereum utilities (for Polymarket)
web3>=6.0.0
eth-account>=0.10.0

# Async support
aiohttp>=3.9.0
